+++ changelog.md
@@ -18,7 +18,7 @@ See also http://pvp.haskell.org/faq
   * Switch to trailer parsing mode only after a full block of `NUL`
   * Drop deprecated `emptyIndex` and `finaliseIndex`
   * Extend `FileNameError` with `UnsafeLinkTarget` constructor
-  * Add `CheckSecurityCallback`, `packWith`, `unpackWith`
+  * Add `packAndCheck` and `unpackAndCheck`
   * Generalize `Entries`, `Entry` and `EntryContent` to `GenEntries`, `GenEntry` and `GenEntryContent`
 
 0.5.1.1 Herbert Valerio Riedel <hvr@gnu.org> August 2019